# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: Home Sale Helper API
  version: "0.1.0"
  description: |
    REST API for the Home Sale Helper micro-SaaS.
    Handles listings, AI generation, photo upload, cross-platform publishing, messaging, and subscriptions.
  contact:
    name: Backend Developer
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.homesalehelper.app
    description: Production (Railway)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          description: JWT access token
        refreshToken:
          type: string
        expiresAt:
          type: integer
          description: Unix timestamp of token expiry
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        listingId:
          type: string
          format: uuid
        platform:
          type: string
          enum: [EBAY, FACEBOOK, POSHMARK, OFFERUP]
        buyerName:
          type: string
          nullable: true
        buyerId:
          type: string
          nullable: true
        content:
          type: string
        direction:
          type: string
          enum: [INBOUND, OUTBOUND]
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /auth/signup:
    post:
      summary: Create account and return JWT
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error or email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      summary: Authenticate and return JWT
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/refresh:
    post:
      summary: Refresh access token
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refreshToken:
                    type: string
                  expiresAt:
                    type: integer
        "401":
          description: Invalid or expired refresh token

  # ── Listings ──────────────────────────────────────────────────────────────
...
  /photos/upload:
    post:
      summary: Upload photos to Supabase Storage (multipart, max 10 files, 20MB each)
      tags: [Photos]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Uploaded photo URLs
          content:
            application/json:
              schema:
                type: object
                properties:
                  urls:
                    type: array
                    items:
                      type: string
                      format: uri
        "400":
          description: No files or unsupported type
  /photos:
    delete:
      summary: Delete photos from Supabase Storage
      tags: [Photos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [urls]
              properties:
                urls:
                  type: array
                  items:
                    type: string
                    format: uri
      responses:
        "204":
          description: Deleted

  # ── AI ────────────────────────────────────────────────────────────────────
...
  /messages:
    get:
      summary: List inbox messages (filterable by platform, listing, unread)
      tags: [Messages]
      parameters:
        - name: unread
          in: query
          schema:
            type: string
            enum: ["true", "false"]
        - name: platform
          in: query
          schema:
            type: string
        - name: listingId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      summary: Record an inbound buyer message
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listingId, platform, content]
              properties:
                listingId:
                  type: string
                  format: uuid
                platform:
                  type: string
                  enum: [EBAY, FACEBOOK, POSHMARK, OFFERUP]
                buyerName:
                  type: string
                buyerId:
                  type: string
                content:
                  type: string
                externalId:
                  type: string
      responses:
        "201":
          description: Message recorded
...
  /messages/listing/{listingId}:
    parameters:
      - name: listingId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get all messages for a listing and mark them read
      tags: [Messages]
      responses:
        "200":
          description: Messages
  /messages/{id}/read:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      summary: Mark a message as read
      tags: [Messages]
      responses:
        "200":
          description: Updated message

  # ── Stripe ────────────────────────────────────────────────────────────────
...